+++ Proyecto/Apis/publicacionesApi/src/main/java/com/example/publicacionesApi/Service/ComentarioService.java
@@ -5,6 +5,8 @@ import java.util.List;
 
 public interface ComentarioService {
     Comentario guardarComentario(Comentario comentario);
-    List<Comentario> obtenerComentariosPorPublicacion(Integer idPublicacion); // Cambiado Long a Integer
-    void eliminarComentario(Long idComentario);
+    List<Comentario> obtenerComentariosPorPublicacion(Integer idPublicacion); 
+    
+    // CORRECCIÓN: Se cambió Long a Integer para coincidir con la llave primaria
+    void eliminarComentario(Integer idComentario);
 }
