+++ Proyecto/Apis/publicacionesApi/src/main/java/com/example/publicacionesApi/Service/RespuestaComentarioService.java
@@ -5,6 +5,8 @@ import java.util.List;
 
 public interface RespuestaComentarioService {
     RespuestaComentario guardarRespuesta(RespuestaComentario respuesta);
-    List<RespuestaComentario> obtenerRespuestasPorComentario(Integer idComentario); // Cambiado Long a Integer
-    void eliminarRespuesta(Long idRespuesta);
+    List<RespuestaComentario> obtenerRespuestasPorComentario(Integer idComentario); 
+    
+    // CORRECCIÓN: Se cambió Long a Integer para coincidir con la llave primaria
+    void eliminarRespuesta(Integer idRespuesta);
 }
